+++ playgrounds/content-v2/nuxt.config.ts
@@ -16,7 +16,4 @@ export default defineNuxtConfig({
     compatibilityVersion: 4,
   },
   compatibilityDate: '2025-01-01',
-  nuxtToc: {
-    depth: 2,
-  },
 })